UI(utils/detection): reordered output in HTML output #1055
